stdc -> libc
