+++ include/libc/stdarg.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,15 +29,14 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _JINUE_STDC_STDLIB_H_
-#define _JINUE_STDC_STDLIB_H_
+#ifndef _JINUE_LIBC_STDARG_H
+#define _JINUE_LIBC_STDARG_H
 
-#include <stddef.h>
+typedef    __builtin_va_list    va_list;
 
-#define EXIT_SUCCESS 0
-
-#define EXIT_FAILURE 1
-
-char *getenv(const char *name);
+#define va_start(ap, parmN)     __builtin_va_start((ap), (parmN))
+#define va_arg                  __builtin_va_arg
+#define va_end                  __builtin_va_end
+#define va_copy(dest, src)      __builtin_va_copy((dest), (src))
 
 #endif
