+++ include/libc/stdlib.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,33 +29,15 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _JINUE_STDC_STDINT_H
-#define _JINUE_STDC_STDINT_H
+#ifndef _JINUE_LIBC_STDLIB_H_
+#define _JINUE_LIBC_STDLIB_H_
 
-#define INT64_C(x)              (x##LL)
+#include <stddef.h>
 
-#define UINT64_C(x)             (x##ULL)
+#define EXIT_SUCCESS 0
 
-typedef signed char             int8_t;
+#define EXIT_FAILURE 1
 
-typedef short int               int16_t;
-
-typedef int                     int32_t;
-
-typedef long long int           int64_t;
-
-
-typedef unsigned char           uint8_t;
-
-typedef unsigned short int      uint16_t;
-
-typedef unsigned int            uint32_t;
-
-typedef unsigned long long int  uint64_t;
-
-
-typedef int                     intptr_t;
-
-typedef unsigned int            uintptr_t;
+char *getenv(const char *name);
 
 #endif
